# GH-102613: Improve performance of `pathlib.Path.rglob()`